refactor: extract form from record page
